% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{archivedir}
\alias{archivedir}
\alias{archivedir,GRANRepository-method}
\title{archivedir
Return the full path to the archive directory for the final repository
deployment.}
\usage{
archivedir(repo)

\S4method{archivedir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The full path to the archive directory where the archived packages
will be deployed to
}
\description{
archivedir
Return the full path to the archive directory for the final repository
deployment.
}
