% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contrib.url-methods.R
\name{available.packages}
\alias{available.packages}
\title{available.packages
A generic for available.packages and a method for GRANRepository objects}
\usage{
available.packages(contriburl = contrib.url(getOption("repos"), type), method,
  fields = NULL, type = getOption("pkgType"), filters = NULL)
}
\description{
available.packages
A generic for available.packages and a method for GRANRepository objects
}

