% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib.url-methods.R
\docType{methods}
\name{available.packages}
\alias{available.packages}
\alias{available.packages,ANY-method}
\alias{available.packages,ANY}
\alias{available.packages,GRANRepository-method}
\alias{available.packages,GRANRepository}
\title{available.packages
A generic for available.packages and a method for GRANRepository objects}
\usage{
available.packages(contriburl, method, fields = NULL,
  type = getOption("pkgType"), filters = NULL, repos = NULL)

\S4method{available.packages}{ANY}(contriburl, method, fields = NULL,
  type = getOption("pkgType"), filters = NULL, repos = NULL)

\S4method{available.packages}{GRANRepository}(contriburl, method,
  fields = NULL, type = getOption("pkgType"), filters = NULL,
  repos = NULL)
}
\arguments{
\item{contriburl}{The repository or contrib url}

\item{method}{See base documentation}

\item{fields}{See base documentation}

\item{type}{The type of packages to query}

\item{filters}{See base documetnation}

\item{repos}{Character string for the repository to query. GRANRepository
objects should be passed to the contriburl argument.}
}
\description{
available.packages
A generic for available.packages and a method for GRANRepository objects
}
