% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{backup_archive}
\alias{backup_archive}
\alias{backup_archive,GRANRepository-method}
\alias{backup_archive,}
\alias{GRANRepository-method}
\title{backup_archive
Return path where packages are backed up by default when clearing the repo}
\usage{
backup_archive(repo)

\S4method{backup_archive}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
Directory where packages are backed up by default when clearing repo
}
\description{
backup_archive
Return path where packages are backed up by default when clearing the repo
}
