% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{buildBadge}
\alias{buildBadge}
\title{Create Badges for build status}
\usage{
buildBadge(status, pkg_name)
}
\arguments{
\item{status}{The build status for the package}

\item{pkg_name}{Name of the package}
}
\value{
Badge href tag
}
\description{
Create Badges for build status
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
