% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBranchesInRepo.R
\name{buildBranchesInRepo}
\alias{buildBranchesInRepo}
\title{Build SVN Checkouts Into Repository Directory}
\usage{
buildBranchesInRepo(repo, cores = 1, temp = FALSE, incremental = TRUE,
  manifest = manifest_df(repo))
}
\arguments{
\item{repo}{a GRANRepository object}

\item{cores}{number of cores to use during build process. defaults to 1}

\item{temp}{logical. whether we are building the temp or final version of the repository}

\item{incremental}{logical. whether packages should only be rebuilt if the version number has increased. Default is TRUE}

\item{manifest}{data.frame containing a GRAN manifest of pkgs to build. Defaults to the full manifest associated with repo}
}
\value{
a list with success and fail elements containing the directories which succeeded and failed the build
}
\description{
create the tarballs in the new repo from the svn branch locs
}
\author{
Cory Barr, Gabriel Becker
}

