% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReport.R
\name{buildReport}
\alias{buildReport}
\title{buildReport}
\usage{
buildReport(repo, theme = "bootstrap",
  reportfile = file.path(destination(repo), "buildreport.html"),
  riskrpt = FALSE, jsonrpt = TRUE, splashname = "index.html")
}
\arguments{
\item{repo}{A GRANRepository object}

\item{theme}{CSS+JS theme. bootstrap, foundation, semanticui or jqueryui}

\item{reportfile}{File path of the HTML report}

\item{riskrpt}{Whether to build the risk report}

\item{jsonrpt}{Whether to create a JSON version of the build report}

\item{splashname}{Filename for the package HTML splash page}
}
\value{
None
}
\description{
Create a build report for a repository reflecting the latest build
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
