% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{buildReportURL}
\alias{buildReportURL}
\title{Constructs the gRAN build report URL}
\usage{
buildReportURL(repo)
}
\arguments{
\item{repo}{A gRAN repo object}
}
\description{
Constructs the gRAN build report URL
}
\note{
This function is not intended for direct use by the end user.
}
