% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{check_result_dir}
\alias{check_result_dir}
\alias{check_result_dir,GRANRepository-method}
\title{check_result_dir
Return the path where check results for packages will be deployed for use in
the build report.}
\usage{
check_result_dir(repo)

\S4method{check_result_dir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The directory where check results should be deployed for use in the
build report
}
\description{
check_result_dir
Return the path where check results for packages will be deployed for use in
the build report.
}
