% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{checkAndFixLibLoc}
\alias{checkAndFixLibLoc}
\title{Protect against binary incompatibility in R versions (3.4- <-> 3.5+)}
\usage{
checkAndFixLibLoc(repo)
}
\arguments{
\item{repo}{GRANRepository being built}
}
\value{
\code{repo}, after clearing the temporary library location
if packages in it were built using a different R version
}
\description{
Protect against binary incompatibility in R versions (3.4- <-> 3.5+)
}
