% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{checkout_dir}
\alias{checkout_dir}
\alias{checkout_dir,GRANRepository-method}
\alias{checkout_dir,NULL-method}
\alias{checkout_dir,NULL}
\title{checkout_dir
Return the directory that package sources will be checked-out into
for use in the build process}
\usage{
checkout_dir(repo)

\S4method{checkout_dir}{GRANRepository}(repo)

\S4method{checkout_dir}{`NULL`}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
For destination, the full path to the contrib directory the packages
will be deployed to
}
\description{
checkout_dir
Return the directory that package sources will be checked-out into
for use in the build process
}
