% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clear.R
\name{clear_temp_fils}
\alias{clear_repo}
\alias{clear_temp_fils}
\title{Clear packages and temporary files  from repo build process}
\usage{
clear_temp_fils(repo, checkout = FALSE)

clear_repo(repo, archivedir = NULL)
}
\arguments{
\item{repo}{GRANRepository - The repository to clean}

\item{checkout}{logical - Should the checkouts of packages also be cleared.
Generally this is not necessary (default: FALSE)}

\item{archivedir}{}
}
\value{
logical scalar indicating full success (TRUE) or one or more failures (FALSE). The repository log will contain more detailed information about failures.
}
\description{
These are convenience functions which clears the intermediate
files generated during the build process. This is important
when, e.g., building a repository for the first time with
a new version of R.
}
\details{
\code{clear_repo} removes packages deployed into the destination repository
and updates the PACKAGES and PACKAGES.gz files. \code{clear_temp_fils} clears
intermediate files from the library location used during building, the temporary
repository, the package staging area, and the store of install- and check-results.
}
\author{
Gabriel Becker
}

