% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contrib.url-methods.R
\name{contrib.url}
\alias{contrib.url}
\title{contrib.url
A generic for contrib.url so that available.packages et al can
interact with GRANRepository objects.}
\usage{
contrib.url(repos, type = getOption("pkgType"))
}
\description{
contrib.url
A generic for contrib.url so that available.packages et al can
interact with GRANRepository objects.
}

