% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{coverage_report_dir}
\alias{coverage_report_dir}
\alias{coverage_report_dir,GRANRepository-method}
\title{coverage_report_dir
Return the path where test coverage reports for packages will be deployed
for use in the build report.}
\usage{
coverage_report_dir(repo)

\S4method{coverage_report_dir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The directory where test coverage results should be deployed for use
in the build report
}
\description{
coverage_report_dir
Return the path where test coverage reports for packages will be deployed
for use in the build report.
}
