% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{createHyperlink}
\alias{createHyperlink}
\title{Creates a href tag}
\usage{
createHyperlink(string, label = "")
}
\arguments{
\item{string}{URL string}

\item{label}{Label for href}
}
\value{
href tag
}
\description{
Creates a href tag
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
