% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgJSON.R
\name{createJSON}
\alias{createJSON}
\title{Create JSON representation of package information}
\usage{
createJSON(repo, pkg_name, descr_df, scm_df, docdir, suffix = paste0("_",
  descr_df$Version, ".json"))
}
\arguments{
\item{repo}{A GRAN repo object}

\item{pkg_name}{Name of the GRAN package}

\item{descr_df}{data.frame representation of DESCRIPTION file}

\item{scm_df}{data.frame representation of GRAN manifest object}

\item{docdir}{Directory where the JSON doc will be written}

\item{suffix}{Suffix for the JSON file}
}
\value{
None. Write JSON file to disk
}
\description{
Create JSON representation of package information
}
\seealso{
\code{\link{manifest_df}} for generating scm_df and
   \code{\link{generateDescInfo}} for generating descr_df.
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
