% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMeta.R
\name{createMeta}
\alias{createMeta}
\title{Create package metadata files}
\usage{
createMeta(repo, repodest = destination(repo), metadir = metadatadir(repo),
  archive_dir = archivedir(repo), serialize_json = FALSE)
}
\arguments{
\item{repo}{A GRAN repo object}

\item{repodest}{The repo destination
(something that looks like BASE_REPO_DIR/src/contrib)}

\item{metadir}{The directory containing metadata files}

\item{archive_dir}{Directory containing package archive}

\item{serialize_json}{Serialize the RDS metadata files as JSON}
}
\description{
Create package metadata files
}
\author{
Dinakar Kulkarni
}
