% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSticker.R
\name{createSticker}
\alias{createSticker}
\title{Create hex stickers for packages}
\usage{
createSticker(pkgName, destination)
}
\arguments{
\item{pkgName}{Name of the package}

\item{destination}{Directory where the sticker will be saved}
}
\value{
None
}
\description{
Create hex stickers for packages
}
