% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{createURL}
\alias{createURL}
\title{Creates a href tag}
\usage{
createURL(string, label = "")
}
\arguments{
\item{string}{URL string}

\item{label}{Label for href}
}
\value{
href tag
}
\description{
Creates a href tag
}
