% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{deltaDF}
\alias{deltaDF}
\title{Returns the difference between 2 data frames}
\usage{
deltaDF(new_df, old_df)
}
\arguments{
\item{new_df}{The new dataframe which you want to compare}

\item{old_df}{An older dataframe of the same structure}
}
\value{
Differences as a dataframe of the same structure
}
\description{
Returns the difference between 2 data frames
}
\note{
This function is not intended for direct use by the end user.
}
\seealso{
\code{\link[dplyr]{anti_join}}
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
