% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2JSON.R
\name{desc2JSON}
\alias{desc2JSON}
\title{Build JSON of DESCRIPTION file}
\usage{
desc2JSON(repo, suffix = "-description.json")
}
\arguments{
\item{repo}{A GRAN repo object}

\item{suffix}{String that is appended to the end of the JSON file}
}
\value{
None
}
\description{
Create a JSON version of the package's DESCRIPTION file. Useful for
interfacing with other applications.
}
