% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{dest_base}
\alias{dest_base}
\alias{dest_base,GRANRepository-method}
\title{dest_base
Return the full path to the contrib directory for the final repository
deployment.}
\usage{
dest_base(repo)

\S4method{dest_base}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
For dest_base, the base directory the repository will reside in
}
\description{
dest_base
Return the full path to the contrib directory for the final repository
deployment.
}
