% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{destination}
\alias{destination}
\alias{destination,GRANRepository-method}
\title{destination
Return the full path to the contrib directory for the final repository
deployment.}
\usage{
destination(repo)

\S4method{destination}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
For destination, the full path to the contrib directory the packages
will be deployed to
}
\description{
destination
Return the full path to the contrib directory for the final repository
deployment.
}
