% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{determinePkgURL}
\alias{determinePkgURL}
\title{Generate package's external URL after validation info as data.frame}
\usage{
determinePkgURL(pkg_name)
}
\arguments{
\item{pkg_name}{The name of the package (string)}
}
\value{
Package external URL
}
\description{
Generate package's external URL after validation info as data.frame
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
