% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{email_notify}
\alias{email_notify}
\alias{email_notify,GRANRepository-method}
\title{email_notify
Should emails be sent for build failure notifications?}
\usage{
email_notify(repo)

\S4method{email_notify}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
logical
}
\description{
email_notify
Should emails be sent for build failure notifications?
}
