% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{email_options}
\alias{email_options}
\alias{email_options,GRANRepository-method}
\title{email_options
Email options for sending build failure notifications}
\usage{
email_options(repo)

\S4method{email_options}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
A list containing the email options
}
\description{
email_options
Email options for sending build failure notifications
}
