% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{emailNotifier}
\alias{emailNotifier}
\title{Send email notifications to maintainers whose builds failed}
\usage{
emailNotifier(repo_name = "", mailControl = list(smtpServer = "localhost"),
  sender = "resgran@gene.com", attachments = NULL,
  logfile = "gRANfailureMail.log", log_level = "DEBUG")
}
\arguments{
\item{repo_name}{gRAN repo_name could be <blank>|devel|current|stable|test}

\item{mailControl}{List object containing SMTP server info}

\item{sender}{Sender's email ID as a string}

\item{attachments}{Files with full paths, as an array}

\item{logfile}{Get or set the logfile for a logger object}

\item{log_level}{Set or get the priority level for a logger object.}
}
\value{
None
}
\description{
Send email notifications to maintainers whose builds failed
}
\note{
This function is not intended for direct use by the end user.
}
\seealso{
\code{\link{getFailureInfo}} for creating failed pkg manifests,
         \code{\link[log4r]{create.logger}} for logger objects,
         \code{\link{sendMail}} for sending emails,
         \code{\link{deltaDF}} for difference between 2 dataframes,
         \code{\link{notifyManifest}} for manifest-based emails
}
\author{
Dinakar Kulkarni, \email{kulkarni.dinakar@gene.com}
}
