% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{emailTag}
\alias{emailTag}
\title{Create a mailto tag for email IDs}
\usage{
emailTag(item)
}
\arguments{
\item{item}{A string with the email ID}
}
\value{
href tag
}
\description{
Create a mailto tag for email IDs
}
