% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{encode_string}
\alias{encode_string}
\title{Convert string to numeric representation}
\usage{
encode_string(x)
}
\arguments{
\item{x}{String}
}
\value{
Numeric representation of string
}
\description{
Convert string to numeric representation
}
\note{
This function is not intended for direct use by the end user.
}
