% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{errlogfile}
\alias{errlogfile}
\alias{errlogfile,GRANRepository-method}
\title{Log file location of a GRAN (sub) repository}
\usage{
errlogfile(repo)

\S4method{errlogfile}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
file location of the errors-only logfile
}
\description{
errlogfile
Retrieve the path to the errors-only logfile for a GRAN repository
}
\author{
Gabriel Becker
}

