% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{extPkgURL}
\alias{extPkgURL}
\title{Wrapper for determinePkgURL}
\usage{
extPkgURL(desc_field, as_string = TRUE)
}
\arguments{
\item{desc_field}{The DESCRIPTION field which has mutiple package names}

\item{as_string}{Return as string}
}
\value{
Package external URL
}
\description{
Wrapper for determinePkgURL
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
