% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{generateDescInfo}
\alias{generateDescInfo}
\title{Converts a DESCRIPTION file to a data.frame}
\usage{
generateDescInfo(pkg_path, encoding = "")
}
\arguments{
\item{pkg_path}{The path preceding the location of the DESCRIPTION file}

\item{encoding}{If there is an Encoding field, to what encoding should
re-encoding be attempted? The other values are as used by iconv, so the
default "" indicates the encoding of the current locale}
}
\value{
If a DESCRIPTION file for the given package is found and can
    successfully be read, this function returns a data.frame ontaining
    of the fields as headers and the tags as rows
}
\description{
Converts a DESCRIPTION file to a data.frame
}
