% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{getFailureInfo}
\alias{getFailureInfo}
\title{Creates a dataframe containing information regarding packages that had a
failed status}
\usage{
getFailureInfo(repo)
}
\arguments{
\item{repo}{A GRAN repo object}
}
\value{
Dataframe containing package info that failed
}
\description{
Creates a dataframe containing information regarding packages that had a
failed status
}
\note{
This function is not intended for direct use by the end user.
}
\seealso{
\code{\link{repo_results}} for repo results as a dataframe
}
