% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOS}
\alias{getOS}
\title{Get the OS Type}
\usage{
getOS()
}
\value{
OS Type
}
\description{
Get the OS Type
}
\note{
This function is not intended for direct use by the end user.
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
