% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GRAN}
\alias{GRAN}
\title{Install packages from GRAN}
\usage{
GRAN(package, repos = c(BiocInstaller::biocinstallRepos(),
  "http://R-Forge.R-project.org"), type = if (onTheMac()) "both" else
  "source", ..., gRAN = GRANRepo$repo)
}
\arguments{
\item{package}{character vector of packages to install from GRAN}

\item{repos}{a list of repos to search in addition to the repos GRAN
searches by default}

\item{type}{The package type; defaults to 'both' on the Mac so
that source packages are attempted.}

\item{...}{Additional parameters that will be passed to}

\item{gRAN}{GRANRepository object to use. Defaults to the default
repository for the current installation, or NULL if no default repository
exists.}
}
\value{
TRUE if successful
}
\description{
Install packages in GRAN along with their dependencies
}
\details{
Installs packags in GRAN along with their dependencies
}
\author{
Cory Barr and Gabriel Becker
}

