% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{repo_name}
\alias{check_note_ok}
\alias{check_note_ok,GRANRepository}
\alias{check_note_ok,GRANRepository-method}
\alias{check_test_on}
\alias{check_test_on,GRANRepository}
\alias{check_test_on,GRANRepository-method}
\alias{check_test_on,RepoBuildParam}
\alias{check_test_on,RepoBuildParam-method}
\alias{check_warn_ok}
\alias{check_warn_ok,GRANRepository}
\alias{check_warn_ok,GRANRepository-method}
\alias{extra_fun}
\alias{extra_fun,GRANRepository-method}
\alias{install_test_on}
\alias{install_test_on,GRANRepository}
\alias{install_test_on,GRANRepository-method}
\alias{install_test_on,RepoBuildParam}
\alias{install_test_on,RepoBuildParam-method}
\alias{logfun,GRANRepository}
\alias{logfun,GRANRepository-method}
\alias{logfun<-,GRANRepository}
\alias{logfun<-,GRANRepository-method}
\alias{repo_name}
\alias{repo_name,GRANRepository}
\alias{repo_name,GRANRepository-method}
\alias{sh_init_script,GRANRepository}
\alias{sh_init_script,GRANRepository-method}
\alias{sh_init_script<-,GRANRepository}
\alias{sh_init_script<-,GRANRepository-method}
\alias{suspended_pkgs}
\alias{suspended_pkgs,GRANRepository}
\alias{suspended_pkgs,GRANRepository-method}
\alias{suspended_pkgs<-}
\alias{suspended_pkgs<-,GRANRepository}
\alias{suspended_pkgs<-,GRANRepository-method}
\alias{temp_repo}
\alias{temp_repo,GRANRepository}
\alias{temp_repo,GRANRepository-method}
\title{Get or set individual parameters on a GRANRepository object}
\usage{
repo_name(x)

\S4method{repo_name}{GRANRepository}(x)

temp_repo(x)

\S4method{temp_repo}{GRANRepository}(x)

check_warn_ok(x)

\S4method{check_warn_ok}{GRANRepository}(x)

check_note_ok(x)

\S4method{check_note_ok}{GRANRepository}(x)

suspended_pkgs(x)

\S4method{suspended_pkgs}{GRANRepository}(x)

suspended_pkgs(x) <- value

\S4method{suspended_pkgs}{GRANRepository}(x) <- value

\S4method{sh_init_script}{GRANRepository}(x)

\S4method{sh_init_script}{GRANRepository}(x) <- value

extra_fun(x)

\S4method{extra_fun}{GRANRepository}(x)

check_test_on(x)

\S4method{check_test_on}{RepoBuildParam}(x)

\S4method{check_test_on}{GRANRepository}(x)

install_test_on(x)

\S4method{install_test_on}{RepoBuildParam}(x)

\S4method{install_test_on}{GRANRepository}(x)

\S4method{logfun}{GRANRepository}(x)

\S4method{logfun}{GRANRepository}(x) <- value
}
\arguments{
\item{x}{A GRANRepository object}

\item{value}{The new parameter value}
}
\description{
Get or set individual parameters on a GRANRepository object
}
\details{
These functions get or set individual repository build parameters
on a GRANRepository object.
}
\seealso{
\code{\link{RepoBuildParam}}
}

