% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GRANRepo}
\alias{GRANRepo}
\title{GRANRepo}
\format{\preformatted{<environment: 0x17027868> 
}}
\usage{
GRANRepo
}
\description{
An environment populated at load time with a GRANRepository object in GRANRepo$repo
}
\keyword{datasets}

