% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GRANRepository}
\alias{GRANRepository}
\title{GRANRepository}
\usage{
GRANRepository(basedir, subrepoName = "stable", rversion = "R",
  tempRepo = file.path(basedir, subrepoName, "tmprepo"),
  tempCheckout = file.path(basedir, "tmpcheckout"),
  errlog = file.path(basedir, subrepoName, paste0("GRAN-errors-", subrepoName,
  "-", Sys.Date(), ".log")), logfile = file.path(basedir, subrepoName,
  paste0("GRAN-log-", subrepoName, "-", Sys.Date(), ".log")),
  checkNoteOk = TRUE, checkWarnOk = TRUE, tempLibLoc = file.path(basedir,
  subrepoName, "LibLoc"), extraFun = function(...) NULL, destination,
  auth = "", manifest = emptyManifest, dest_url, shell_init = "")
}
\arguments{
\item{basedir}{The base directory. By default the temporary repository,
temporary install library, and package staging area will be located in
<basedir>/<subrepoName>/, while the  temporary source checkout will be in t
he basedir itself.}

\item{subrepoName}{The name of the repository, e.g. stable or devel}

\item{rversion}{The R executable to use when invoking R CMDs and package
testing code}

\item{tempRepo}{Location to create the temporary repository}

\item{tempCheckout}{Location to create temporary checkouts/copies of package
source code}

\item{errlog}{The file to append error output to during the building and
testing processes}

\item{logfile}{The file to append summary log information to during building
and testing}

\item{checkNoteOk}{logical. Whether packages that raise notes during
R CMD check should be considered to have passed}

\item{checkWarnOk}{logical. Whether packages that raise warnings during
R CMD check should be considered to have passed}

\item{tempLibLoc}{Location to create the temporary installed package library
for use during the testing process}

\item{extraFun}{currently ignored}

\item{destination}{Base location (not including repository name) of the
final repository to be built}

\item{auth}{character. Authentication information required to add packages
to the manifest.}

\item{manifest}{data.frame. The manifest of package information associated
with this repository. Defaults to an empty manifest.}

\item{dest_url}{The base URL the destination directory corresponds to. The
subrepository name will be appended to this to generate the URL used when
installing from the repository.}

\item{shell_init}{An optional shell script to source before invoking system
commands, e.g. a bashrc file. Ignored if "" or not specified.}
}
\description{
A constructor for the \code{GRANRepository} class of S4 objects representing
individual repositories
}

