% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.R
\docType{methods}
\name{install_packages,character,GRANRepository-method}
\alias{install_packages,character,GRANRepository-method}
\title{install_packages method for GRANRepository objects}
\usage{
\S4method{install_packages}{character,GRANRepository}(pkgs, repos,
  versions = NULL, verbose = FALSE, ...)
}
\arguments{
\item{pkgs}{Character vector of package names to install}

\item{repos}{A GRANRepository}

\item{versions}{Ignored}

\item{verbose}{Whether verbose/debugging information should be printed}

\item{\dots}{Passed (eventually) to install.packages}
}
\description{
install_packages method for GRANRepository objects
}
\details{
This method calls directly down to the character,character
method of install_packages in switchr
}
