% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{install_result_dir}
\alias{install_result_dir}
\alias{install_result_dir,GRANRepository-method}
\title{install_result_dir
Return the path where instal results for packages will be deployed for use in
the build report.}
\usage{
install_result_dir(repo)

\S4method{install_result_dir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The directory where install results should be deployed for use in the
build report
}
\description{
install_result_dir
Return the path where instal results for packages will be deployed for use in
the build report.
}
