% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installPkgVersion}
\alias{installPkgVersion}
\title{installPkgVersion}
\usage{
installPkgVersion(name, version, repo = NULL, dir = if (is.null(repo))
  tempdir() else notrack(repo), libloc = .libPaths()[1], ...)
}
\arguments{
\item{name}{package name}

\item{version}{package version string}

\item{repo}{(optional) GRANRepository object to search}

\item{dir}{directory to download package into}

\item{libloc}{library to install into. Defaults to first
element of \code{.libPaths()}}
}
\value{
The full path to the downloaded file , or NULL if unable to
locate the package
}
\description{
Install an exact version of a single package
}
\note{
This function does *NOT* find old versions of the package's
dependencies. It is safer to create a full (temporary) sessionRepository
using \code{sessionRepo}
}

