% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{isValidEmail}
\alias{isValidEmail}
\title{Checks whether an email ID is valid}
\usage{
isValidEmail(email_id)
}
\arguments{
\item{email_id}{Email ID as a string}
}
\value{
Boolean
}
\description{
Checks whether an email ID is valid
}
\note{
This function is not intended for direct use by the end user.
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
