% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lazyRepo}
\alias{lazyRepo}
\title{lazyRepo}
\usage{
lazyRepo(pkgs, manifest, versions = rep(NA, times = length(pkgs)),
  dir = tempdir(), rep_path = file.path(dir, "repo"),
  get_suggests = FALSE, verbose = FALSE, scm_auths = list(bioconductor =
  c("readonly", "readonly")))
}
\description{
lazyRepo
}

