% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{LibLoc}
\alias{LibLoc}
\alias{LibLoc,GRANRepository-method}
\title{LibLoc
Return the temporary library location associated with the repository}
\usage{
LibLoc(repo)

\S4method{LibLoc}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the repository specific library location
}
\description{
LibLoc
Return the temporary library location associated with the repository
}

