% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANonGRAN.R
\name{loadRepo}
\alias{loadRepo}
\title{Backwards compatible load utility}
\usage{
loadRepo(filename)
}
\arguments{
\item{filename}{The file to load}
}
\description{
Load a repository serialized to an R code file
}

