% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{location,GRANRepository-method}
\alias{location,GRANREpository-method}
\alias{location,GRANRepository-method}
\title{Retrieve the path to a GRAN (sub) repository}
\usage{
\S4method{location}{GRANRepository}(repo)
}
\description{
Retrieve the path to a GRAN (sub) repository
}

