% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeGRANRepos}
\alias{makeGRANRepos}
\title{Build a set of GRAN repositories}
\usage{
makeGRANRepos(manifest = read.table(file.path(baseDir, subRepos,
  "GRANmanifest.dat")), Rlocs = c(stable = "R", dev = "Rdevel"),
  tmpRepoBase = "./tmpRepos", baseDir = "./GRANRepos",
  tmpCheckout = "./tmpSourceDirs", logfiles = file.path(baseDir, subRepos,
  paste0("GRANLog-", Sys.Date(), ".log")), errfiles = file.path(baseDir,
  subRepos, paste0("GRANErrors-", Sys.Date(), ".log")),
  tempLibLocs = file.path(baseDir, subRepos, "LibLoc"), checkWarnOk = FALSE,
  checkNoteOk = TRUE, cores = 3L, repos = list(NULL), auth = "",
  scm_auths = list(list()), dest_bases = normalizePath2(baseDir),
  dest_urls = paste0("file://", normalizePath2(dest_bases)),
  shell_inits = "")
}
\arguments{
\item{manifest}{a data.frame containing the name, url, type of checkout, subdir, subrepo, branch, and extra. This single manifest should contain information for all repositories to be built. Overridden if repos are specified}

\item{Rlocs}{named character vector of the copies of the R executable to use for each of the subRepos. e.g. c(stable="R", dev = "/path/to/Rdevel/bin/R")}

\item{tmpRepoBase}{the base directory for the temporary intermediate repository}

\item{baseDir}{the base directory to use when building each repository.}

\item{tmpCheckout}{the directory the package sources will be checkedout into from version control before being built for each repository}

\item{logfiles}{character vector indicating the paths that the summary logfile should be written to for each repository}

\item{errfiles}{character vector indicating the paths that the error output logfile should be written to for each repository}

\item{tempLibLocs}{The temporary library location to use for each repository}

\item{checkWarnOk}{logical vector specifying whether packages that pass check with warnings should be included in each resulting respository}

\item{checkNoteOk}{logical vector specifying whether packages that pass check with notes should be included in  each resulting repository}

\item{cores}{integer indicating how many cores should be used to build the repositories}

\item{repos}{A list of existing GRANRepository objects to use when building the repositories.}

\item{auth}{character value indicating authentication required to add to manifest}

\item{scm_auths}{named list. Names are regular expressions which match scm (SVN, git) repositories. Elements are character vectors providing the username and password to checkout from the repository. Default is to use readonly:readonly for bioconductor.}

\item{dest_bases}{Base directory (or directories) to contain each named repository.}

\item{dest_urls}{The URLs where the repositories will be hosted. Defaults to file://<dest_bases>.}

\item{shell_inits}{The path to a file for the shell to source before running
system commands. E.g. a bashrc file.}
}
\value{
A list of GRANRepository objects for the created repositoriesxs
}
\description{
makeGRANRepos
Fire build procoess for a GRAN instance (all subrepositories)
}
\author{
Gabriel Becker, heavily adapted from a previous version by Cory Barr
}
\seealso{
Other buildRepos: \code{\link{makeSingleGRANRepo}}
}

