% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeManifestRow}
\alias{makeManifestRow}
\title{makeManifestRow}
\usage{
makeManifestRow(name = NA, url = NA, type = NA, subrepo = "current",
  branch = NA, subdir = NA, extra = NA, status = NA, building = NA,
  version = NA, lastAttempt = NA, lastAttemptVersion = NA,
  lastAttemptStatus = NA, lastbuilt = NA, lastbuiltversion = NA,
  lastbuiltstatus = NA, buildReason = NA, maintainer = NA,
  suspended = FALSE)
}
\arguments{
\item{name}{name of the package}

\item{url}{location of the package sources}

\item{type}{type of location (svn, git, local, etc)}

\item{subrepo}{name of the repository to build the package in}

\item{branch}{name of the branch to use to build the package}

\item{subdir}{subdirectory to use to build the package}

\item{extra}{currently ignored. extra commands for building or
installing the package}

\item{status}{status of the current build (typically starts as 'ok')}

\item{building}{whether the package is being/should be built in this
building cycle}

\item{version}{Version of the package.}

\item{lastAttempt}{Date of the last attempted build (or NA)}

\item{lastAttemptVersion}{version used in last attempted build (or NA)}

\item{lastAttemptStatus}{result of the last attempted build (or NA)}

\item{lastbuilt}{Date of the last successful build (or NA)}

\item{lastbuiltversion}{version built duirng last successful build (or NA)}

\item{lastbuiltstatus}{result of the last successful build (or NA)}

\item{buildReason}{Reason for building the package. typically NA}

\item{maintainer}{The maintainer of the package}

\item{suspended}{Whether building of the package is suspended (without
removal from the manifest). Defaults to FALSE}
}
\value{
A valid GRAN manifest data.frame
}
\description{
Create one or more rows of manifest
}

