% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeRepo}
\alias{makeRepo}
\title{makeRepo}
\usage{
makeRepo(x, cores = 3L, build_pkgs = NULL,
  scm_auth = list(bioconductor.org = c("readonly", "readonly")), ...)
}
\arguments{
\item{x}{The object containing the information necessary to create the repository}

\item{cores}{The number of cores on the local machine to use during building}

\item{build_pkgs}{The names of the packages to (re) build and test within the
repository. Defaults to \code{NULL} which builds all packages in the manifest}

\item{scm_auth}{A named list containing the information necessary to check
out package sources. The list elements (assumed to be a character vector of
length 2, user then password) are applied when the name is contained in a
package's url}

\item{...}{Additional arguments, typically used for the construction of a
\code{GRANRepository} object if one does not already exist.}
}
\value{
A GRANRepository object which has used to create a repository.
}
\description{
Make a package repository containing a specified set of packages from
various sources
}

