% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestReport.R
\name{manifestReport}
\alias{manifestReport}
\title{manifestReport}
\usage{
manifestReport(repo, theme = "bootstrap",
  reportfile = file.path(destination(repo), "manifest.html"),
  jsonrpt = TRUE)
}
\arguments{
\item{repo}{A GRANRepository object}

\item{theme}{CSS+JS theme. bootstrap, foundation, semanticui or jqueryui}

\item{reportfile}{File path of the HTML report}

\item{jsonrpt}{Whether to create a JSON version of the manifest report}
}
\value{
None
}
\description{
Build a package manifest report for a GRAN repository
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
