% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{notrack}
\alias{notrack}
\alias{notrack,GRANRepository-method}
\title{notrack
Return the directory which stores retreived versions of non-GRAN packages
for use in virtual repositories}
\usage{
notrack(repo)

\S4method{notrack}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the notrack directory
}
\description{
notrack
Return the directory which stores retreived versions of non-GRAN packages
for use in virtual repositories
}

