% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{pkg_doc_dir}
\alias{pkg_doc_dir}
\alias{pkg_doc_dir,GRANRepository-method}
\title{pkg_doc_dir
Return the path where test coverage reports for packages will be deployed
for use in the build report.}
\usage{
pkg_doc_dir(repo)

\S4method{pkg_doc_dir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The directory where test coverage results should be deployed for use
in the build report
}
\description{
pkg_doc_dir
Return the path where test coverage reports for packages will be deployed
for use in the build report.
}
