% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{relatedPkgs}
\alias{relatedPkgs}
\title{Wrapper for dependsOnPkgs}
\usage{
relatedPkgs(pkg_name, relation = "LinkingTo")
}
\arguments{
\item{pkg_name}{The name of the package (string)}

\item{relation}{What type of reverse dependency?}
}
\value{
String of reverse dependencies
}
\description{
Wrapper for dependsOnPkgs
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
