% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{repobase}
\alias{repobase}
\alias{repobase,GRANRepository-method}
\title{repobase
Generic accessor function to retreive the repo specific subdirectory within the base directory}
\usage{
repobase(repo)

\S4method{repobase}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the repository specific directory
}
\description{
repobase
Generic accessor function to retreive the repo specific subdirectory within the base directory
}
