% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{repoManifest}
\alias{repoManifest}
\alias{repoManifest,GRANRepository-method}
\title{repoManifest
Return the location of the manifest file associated with the repository}
\usage{
repoManifest(repo)

\S4method{repoManifest}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the manifest file
}
\description{
repoManifest
Return the location of the manifest file associated with the repository
}

