% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{repo_results}
\alias{repo_results}
\alias{repo_results,GRANRepository-method}
\alias{repo_results,GRANRepository}
\alias{repo_results<-}
\alias{repo_results<-,GRANRepository-method}
\alias{repo_results<-,GRANRepository}
\title{Repository build results}
\usage{
repo_results(x)

\S4method{repo_results}{GRANRepository}(x)

repo_results(x) <- value

\S4method{repo_results}{GRANRepository}(x) <- value
}
\arguments{
\item{x}{A GRANRepository object}

\item{value}{The new results data.frame}
}
\value{
A data.frame of build results
}
\description{
Repository build results
}
