% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RepoToList}
\alias{RepoToList}
\title{Transform a GRANRepository object into a list}
\usage{
RepoToList(repo)
}
\arguments{
\item{repo}{repository}
}
\value{
a list suitable for use with RepoFromList
}
\description{
Utility to transform a GRANRepository object into a list
so that repos saved using GRANBase can be loaded by GRAN
without requiring GRANBase
}

