% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{reversals}
\alias{reversals}
\title{Generate reverse dependency info as data.frame}
\usage{
reversals(pkg_name)
}
\arguments{
\item{pkg_name}{The name of the package (string)}
}
\value{
data.frame containing package reverse dependency info
}
\description{
Generate reverse dependency info as data.frame
}
