% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GRANonGRAN.R
\name{saveRepo}
\alias{saveRepo}
\title{Backwards compatible save utility}
\usage{
saveRepo(repo, filename)
}
\arguments{
\item{repo}{The GRANRepository object to save}

\item{filename}{The destination file}
}
\description{
serialize a repository to a file so that it does not require GRANBase
to load
}

