% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{sortDelimitedString}
\alias{sortDelimitedString}
\title{Alphabetically sort delimited strings}
\usage{
sortDelimitedString(string, delimiter = ",", ...)
}
\arguments{
\item{string}{A string}

\item{delimiter}{A delimiter that separates contents of string}

\item{...}{Arguments that you want to pass to base::sort()}
}
\value{
String that is alphabetically sorted
}
\description{
Alphabetically sort delimited strings
}
\note{
This function is not intended for direct use by the end user.
}
