% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{string2list}
\alias{string2list}
\title{Converts delimited string to list}
\usage{
string2list(string, separator = ",")
}
\arguments{
\item{string}{Input string}

\item{separator}{The delimiter}
}
\value{
Processed list
}
\description{
Converts delimited string to list
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
