% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{temp_lib}
\alias{temp_lib}
\alias{temp_lib,GRANRepository-method}
\alias{temp_lib,GRANRepository}
\title{temporary library}
\usage{
temp_lib(repo)

\S4method{temp_lib}{GRANRepository}(repo)
}
\arguments{
\item{repo}{A GRANRepository object}
}
\description{
temporary library
}
