% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sessionRepo}
\alias{getPkgVersions}
\alias{getSessionPackages}
\alias{makeVirtualRepo}
\alias{sessionRepo}
\title{Create a virtual repository containing only the specified package versions versions


Packages are located via the \code{getPkgVersions} function, which will look in the following places:
\enumerate{
\item{The \code{repo} repository and associated notrack directory}
\item{The current CRAN repository}
\item{The CRAN archives of previous source packages}
\item{The current Bioconductor repository}
\item{The Bioconductor SVN history}
\item{The SCM (SVN/git) history for a GRAN package}
}

When found, package versions not already in the GRAN repository proper or notrack directory are built into the \code{repo}'s associated notrack directory.

The repository is the constructed as a sibling to \code{repo}'s repository using only symbolic links. This allows many virtual repositories to contain the same versions of packages without physical file duplication.}
\usage{
sessionRepo(sinfo = sessionInfo(), repo_dir, doi = NULL, dir, name = NULL,
  replace = FALSE, stoponfail = TRUE, GRepo = GRANRepo$repo,
  install = FALSE, libloc = NULL, biocUseSVN = FALSE)

getSessionPackages(sinfo, dir, GRepo = NULL, stoponfail = FALSE,
  biocUseSVN = FALSE)

getPkgVersions(pkgs, dir, GRepo = NULL, stoponfail = FALSE,
  pkgcol = "Package", verscol = "Version", biocUseSVN = FALSE)

makeVirtualRepo(pkgdf, repo_dir, doi, dir, name = NULL, replace = FALSE,
  stoponfail = TRUE, GRepo, install = FALSE, libloc = NULL, Rvers = "",
  pkgcol = "Package", verscol = "Version", biocUseSVN = FALSE)
}
\arguments{
\item{sinfo}{A sessionInfo object or character vector containing the text from printing such an object}

\item{pkgs}{A data.frame  of package versions to locate and/or build}

\item{pkgcol}{The column in \code{pkgs} or \code{pkgsdf} containing the package names}

\item{verscol}{The column in \code{pkgs} or \code{pkgsdf} containing the package versions}

\item{pkgdf}{A data.frame containing the package names and versions to populate the repository with}

\item{repo_dir}{The base directory to create the virtual repository under.}

\item{doi}{A DOI associated with the session info. If specified when name is NULL, the repository name is set to the doi with "/" replaced with "_".}

\item{dir}{The directory to download/build package tarballs into during the search process}

\item{name}{The name of the repository to create. Defaults to a 32 character hash generated from \code{sinfo}}

\item{replace}{logical. Indicates whether the newly created virtual repository should overwrite any exists virtual repositories of the same name}

\item{stoponfail}{logical. Indicates whether the function should throw an error if it is unable to retreive one or more of the specified package versions. Defaults to \code{TRUE}}

\item{GRepo}{(optional) a \code{GRANRepository} to act as a parent to the
virtual repository. If specified, this is used to: search for pkg versions, determine where to download/build newly located pkg versions, and set the parent directory of the virtual repo.}

\item{install}{should the packages be immediately installed into
\code{libloc}. Defaults to FALSE}

\item{libloc}{If packages are being installed, a library location to
contain only the packages for this set of package versions. In generally this should
*not* be your standard library location.}

\item{Rvers}{The R version to build into the repository structure, if desired. Defaults to no specific version (suitable for src packages).}

\item{pkgcol}{Column in the dataframe that contains package name}

\item{verscol}{Column in the dataframe taht contains package version}

\item{biocUseSVN}{Use SVN instead of git-svn to trawl Bioc repository. This is faster (though not fast) for single use but does NOT amortize future costs.}
}
\value{
for \code{getPkgVersions} and \code{getSessionPackages}, a character vector with the full path to each downloaded/built tar.gz file.

for \code{makeVirtualRepo} and \code{sessionRepo}, the path to the created virtual repository
}
\description{
makeVirtualRepo
}
\details{
Create a virtual repository which contains only the exact packages specified
}
\author{
Gabriel Becker
}

