% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{writeGRANLog}
\alias{writeGRANLog}
\title{writeGRANLog}
\usage{
writeGRANLog(pkg, msg, type = "full", logfile, errfile)
}
\arguments{
\item{pkg}{The name of the package the log is about}

\item{msg}{The log message, collapsed if length>1}

\item{type}{"full", "error", or "both" indicating which log(s) the message
should be written to}

\item{logfile}{The location of the full log file to write/append to}

\item{errfile}{the location of the error log file to write/append to}
}
\description{
Utility function which writes gran logs
}
\note{
This function is not intended for direct use by the end user.
}

