% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contrib.url-methods.R
\docType{methods}
\name{available.packages}
\alias{available.packages}
\alias{available.packages,GRANRepository}
\alias{available.packages,GRANRepository-method}
\title{available.packages
A generic for available.packages and a method for GRANRepository objects}
\usage{
available.packages(contriburl = contrib.url(getOption("repos"), type), method,
  fields = NULL, type = getOption("pkgType"), filters = NULL)


  \S4method{available.packages}{GRANRepository}(contriburl = contrib.url(getOption("repos"),
  type), method, fields = NULL, type = getOption("pkgType"),
  filters = NULL)
}
\arguments{
\item{contriburl}{The repository or contrib url}

\item{method}{See base documentation}

\item{fields}{See base documentation}

\item{type}{The type of packages to query}

\item{filters}{See base documetnation}
}
\description{
available.packages
A generic for available.packages and a method for GRANRepository objects
}

