% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{buildReportURL}
\alias{buildReportURL}
\title{Constructs the gRAN build report URL}
\usage{
buildReportURL(repo_name, url_home = "http://restst.gene.com/gran/")
}
\arguments{
\item{repo_name}{gRAN repo_name could be <blank>|devel|current|stable|test}

\item{url_home}{Base URL for where build reports are hosted}
}
\value{
URL for the gRAN build report
}
\description{
Constructs the gRAN build report URL
}
\note{
This function is not intended for direct use by the end user.
}
