% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{emailNotifier}
\alias{emailNotifier}
\title{Send email notifications to maintainers whose builds failed}
\usage{
emailNotifier(repo, mailControl = list(smtpServer = "localhost"),
  sender = paste0("resgran-d@", system("hostname -d", intern = TRUE)),
  attachments = NULL)
}
\arguments{
\item{repo}{A gRAN repo object}

\item{mailControl}{List object containing SMTP server info}

\item{sender}{Sender's email ID as a string}

\item{attachments}{Files with full paths, as an array}
}
\value{
None
}
\description{
Send email notifications to maintainers whose builds failed
}
\seealso{
\code{\link{getFailureInfo}} for creating failed pkg manifests,
         \code{\link{sendMail}} for sending emails,
         \code{\link{deltaDF}} for difference between 2 dataframes,
         \code{\link{notifyManifest}} for manifest-based emails
}
