% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{emailNotifier}
\alias{emailNotifier}
\title{Send email notifications to maintainers whose builds failed}
\usage{
emailNotifier(repo, mailopts = email_options(repo), attachments = NULL)
}
\arguments{
\item{repo}{A gRAN repo object}

\item{mailopts}{Email options as a list}

\item{attachments}{Files with full paths, as an array}
}
\value{
None
}
\description{
Send email notifications to maintainers whose builds failed
}
\seealso{
\code{\link{getFailureInfo}} for creating failed pkg manifests,
         \code{\link{sendMail}} for sending emails
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
