% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{errlogfile}
\alias{errlogfile}
\alias{errlogfile,GRANRepository-method}
\alias{errlogfile,RepoBuildParam-method}
\title{Log file location of a GRAN (sub) repository}
\usage{
errlogfile(repo)

\S4method{errlogfile}{GRANRepository}(repo)

\S4method{errlogfile}{RepoBuildParam}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
file location of the errors-only logfile
}
\description{
errlogfile
Retrieve the path to the errors-only logfile for a GRAN repository
}
\author{
Gabriel Becker
}
