% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/myGRAN.R
\docType{data}
\name{GRANRepo}
\alias{GRANRepo}
\title{GRANRepo}
\format{\preformatted{<environment: 0x36b3e28> 
}}
\usage{
GRANRepo
}
\description{
An environment populated at load time with a GRANRepository object in GRANRepo$repo
}
\keyword{datasets}

