% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{logfile}
\alias{logfile}
\alias{logfile-method,GRANRepository}
\alias{logfile,GRANRepository-method}
\alias{logfile,RepoBuildParam-method}
\title{Log file location of a GRAN (sub) repository}
\usage{
logfile(repo)

\S4method{logfile}{GRANRepository}(repo)

\S4method{logfile}{RepoBuildParam}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
file location of the full logfile
}
\description{
logfile
}
\details{
Retrieve the path to the full logfile for a GRAN repository
}
\author{
Gabriel Becker
}
