% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GRANRepository-class}
\alias{GRANRepository-class}
\alias{GRANRepository}
\title{GRANRepository}
\usage{
GRANRepository(manifest, results, param = RepoBuildParam(...), ...)
}
\arguments{
\item{manifest}{A PkgManifest object}

\item{results}{A data.frame containing previous build results}

\item{param}{A RepoBuildParam object controlling the location and behavior of
the repository being built}

\item{...}{Passed through to the default value of \code{param}}
}
\description{
A constructor for the \code{GRANRepository} class of S4 objects representing
individual repositories
}
