% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{location}
\alias{location}
\alias{location,GRANREpository-method}
\alias{location,GRANRepository-method}
\alias{location,PkgSource-method}
\title{location generic
Retreive the directory associated with an object}
\usage{
location(repo)

\S4method{location}{GRANRepository}(repo)

\S4method{location}{PkgSource}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
a character containing the associated path
}
\description{
location generic
Retreive the directory associated with an object

Retrieve the path to a GRAN (sub) repository

Retreive the local directory associated with a PkgSource object for a package in a GRAN manifest
}
\author{
Gabriel Becker
}

