% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWinBins.R
\name{makeWinBins}
\alias{makeWinBins}
\title{Make Windows binary packages}
\usage{
makeWinBins(repo, cores = (parallel:::detectCores() - 1),
  virtualstore = file.path(Sys.getenv("LOCALAPPDATA"), "VirtualStore"))
}
\arguments{
\item{repo}{A GRANRepository object}

\item{cores}{Number of cores to use during build process}

\item{virtualstore}{Windows VM directory where built binaries are stored}
}
\value{
None
}
\description{
Create Windows binary builds (only works on Windows machines)
}
\author{
Dinakar Kulkarni
}
