% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestHTML.R
\name{manifestHTML}
\alias{manifestHTML}
\title{manifestHTML}
\usage{
manifestHTML(repo, theme = "bootstrap",
  reportfile = file.path(destination(repo), "buildreport.html"),
  riskrpt = FALSE, jsonrpt = FALSE)
}
\arguments{
\item{repo}{A GRANRepository object}

\item{theme}{CSS+JS theme. bootstrap, foundation, semanticui or jqueryui}

\item{reportfile}{Where the report should be located}

\item{riskrpt}{Whether to build the risk report}

\item{jsonrpt}{Whether to create a JSON version of the build report}
}
\value{
None
}
\description{
Create a build report for a repository reflecting the latest build
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
