% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestHTML.R
\name{manifestHTML}
\alias{manifestHTML}
\title{manifestHTML}
\usage{
manifestHTML(repo)
}
\arguments{
\item{repo}{a GRANRepository object.}
}
\description{
Create a build report for a repository reflecting the latest build
}
