% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{notifyManifest}
\alias{notifyManifest}
\title{Sends email notifications for a given manifest}
\usage{
notifyManifest(manifest, repo, ...)
}
\arguments{
\item{manifest}{A dataframe obtained from getFailureInfo}

\item{repo}{A gRAN repo object}

\item{...}{Additonal arguments to GRANBase::sendMail()}
}
\value{
None
}
\description{
Sends email notifications for a given manifest
}
\note{
This function is not intended for direct use by the end user.
}
\seealso{
\code{\link{sendMail}} for sending emails,
         \code{\link{isValidEmail}} for validating email,
         \code{\link[htmlTable]{htmlTable}} for creating HTML tables
}
