% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{notrack,GRANRepository-method}
\alias{notrack,GRANRepository-method}
\title{notrack
Return the directory which stores retreived versions of non-GRAN packages
for use in virtual repositories}
\usage{
\S4method{notrack}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the notrack directory
}
\description{
notrack
Return the directory which stores retreived versions of non-GRAN packages
for use in virtual repositories
}
