% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{param}
\alias{param}
\alias{param,GRANRepository-method}
\alias{param,GRANRepository}
\alias{param<-}
\alias{param<-,GRANRepository-method}
\alias{param<-,GRANRepository}
\title{Extract parameter object}
\usage{
param(x)

\S4method{param}{GRANRepository}(x)

param(x) <- value

\S4method{param}{GRANRepository}(x) <- value
}
\arguments{
\item{x}{An object with an associated paramater}

\item{value}{A new parameter object}
}
\description{
Extract parameter object
}
