% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgHTML.R
\name{pkgHTML}
\alias{pkgHTML}
\title{Create HTML splash pages for packages}
\usage{
pkgHTML(repo, suffix = "-index.html", theme = "bootstrap")
}
\arguments{
\item{repo}{A gRAN repo object}

\item{suffix}{Append a suffix to the HTML splash page}

\item{theme}{CSS theme. bootstrap, foundation, semanticui or jqueryui}
}
\value{
None
}
\description{
Create HTML splash pages for packages
}
\author{
Dinakar Kulkarni <kulkard2@gene.com>
}
