% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{repo_url}
\alias{repo_url}
\alias{repo_url,GRANRepository-method}
\alias{repo_url,NULL-method}
\alias{repo_url,NULL}
\title{repo_url
Return the url that the repository will be served at. This is the web
address corresponding to repository, suitable for calling contrib.url.  e.g
http://www.website.com/gran/current/}
\usage{
repo_url(repo)

\S4method{repo_url}{GRANRepository}(repo)

\S4method{repo_url}{`NULL`}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
For destination, the full path to the contrib directory the packages
will be deployed to
}
\description{
repo_url
Return the url that the repository will be served at. This is the web
address corresponding to repository, suitable for calling contrib.url.  e.g
http://www.website.com/gran/current/
}
