% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRANonGRAN.R
\name{RepoFromList}
\alias{RepoFromList}
\title{Create a GRANRepository object from a list}
\usage{
RepoFromList(rlist)
}
\arguments{
\item{rlist}{A list with entries that are slot name-value
pairs for a GRANRepository object}
}
\value{
a GRANRepository object
}
\description{
Create a GRANRepository object from a list
}
