% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailNotifier.R
\name{sendMail}
\alias{sendMail}
\title{Wrapper for sendmail, allows for multiple attachments to be sent}
\usage{
sendMail(receiver, subject, body, mailControl = list(smtpServer =
  "localhost"), sender = "my@example.com", attachments = NULL)
}
\arguments{
\item{receiver}{Receiver's email ID as a string. Vector if multiple email IDs}

\item{subject}{Email subject as a string}

\item{body}{Email message body as a string}

\item{mailControl}{List object containing SMTP server info}

\item{sender}{Sender's email ID as a string}

\item{attachments}{Files with full paths as an array}
}
\value{
None
}
\description{
Wrapper for sendmail, allows for multiple attachments to be sent
}
\note{
This function is not intended for direct use by the end user.
}
\seealso{
\code{\link[sendmailR]{sendmail}} for sending simple emails
}
