% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{staging}
\alias{staging}
\alias{staging,GRANRepository-method}
\title{staging
Return the staging directory to be used when building the repository. If the directory does not exist it will be created.}
\usage{
staging(repo)

\S4method{staging}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The path to the repository specific directory
}
\description{
staging
Return the staging directory to be used when building the repository. If the directory does not exist it will be created.
}

